third_party/blink/web_tests/wpt_internal/display-lock/rendersubtree/activation/focus.html

<!doctype HTML>
<html>
<meta charset="utf8">
<title>Display Locking: focus via tab navigation</title>
<link rel="author" title="Rakina Zata Amni" href="mailto:rakina@chromium.org">
<link rel="help" href="https://github.com/WICG/display-locking">
<script src="../resources/utils.js"></script>

<div id=spacer style="height: 3000px"></div>
<div id="container" style ="contain:style layout">
  <div id="focusable" tabIndex="0">
    Focusable div
  </div>
</div>

<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>

<script>
promise_test(() => {
  const acquirePromise = setInvisibleActivatable(container);
  return new Promise((resolve, reject) => {
    const eventPromise = new Promise((resolve, reject) => {
      container.onbeforeactivate = (e) => {
        assert_equals(e.target, container);
        assert_equals(e.activatedElement, focusable);
        resolve();
      };
    });

    acquirePromise.then(() => {
      eventSender.keyDown("Tab", ["shiftKey"]);
      eventPromise.then(() => {
        assert_equals(document.activeElement, focusable);
        resolve();
      });
    });
  });
}, "Activating locked element through tabindex navigation fires beforeactivate, focuses element");
</script>
</html>
